com.github.jackalsin.leetcode.problems/src/test/java/_1401_1450/_1441_Build_an_Array_With_Stack_Operations/SolutionITest.java
package _1401_1450._1441_Build_an_Array_With_Stack_Operations;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.List;
import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 5/9/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int[] target = {1, 3};
    final List<String> expected = List.of(
        "Push", "Push", "Pop", "Push"
    ), actual = solution.buildArray(target, 3);
    assertEquals(expected, actual);
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}